+++ packages/yak_utils/lib/src/classes/union.dart
@@ -1,0 +1,5 @@
+class Union<A, B> {
+  final A a;
+  final B b;
+  const Union(this.a, this.b);
+}
