+++ packages/yak_utils/lib/src/tools/on_error/on_error_delegates.dart
@@ -1,5 +1,4 @@
-import 'package:flutter/material.dart' show debugPrint;
-
+import 'all.dart';
 import 'on_error_types.dart';
 
 class OnErroDoNothing implements OnErrorDelegate {
@@ -8,17 +7,11 @@ class OnErroDoNothing implements OnErrorDelegate {
   void call(Object? p0, StackTrace? p1) {}
 }
 
-class OnErrorPrint implements OnErrorDelegate {
-  const OnErrorPrint();
-  @override
-  void call(Object? p0, StackTrace? p1) => debugPrint('$p0\n$p1');
-}
-
 class RecordError implements OnErrorDelegate {
   RecordError._();
   static final _instance = RecordError._();
   factory RecordError() => _instance;
-  OnError onError = const OnErrorPrint();
+  OnError onError = const OnErroDoNothing();
 
   @override
   void call(Object? p0, StackTrace? p1) => onError(p0, p1);
