+++ packages/yak_utils/lib/src/tools/on_error/on_error_types.dart
@@ -1,4 +1,5 @@
-import 'package:yak_runner/yak_runner.dart';
+import '../../delegates/all.dart';
+import '../../types/all.dart';
 
 typedef OnError = MultiArgFunction2<void, Object?, StackTrace?>;
 typedef OnErrorDelegate = MultiDelegate2<void, Object?, StackTrace?>;
