+++ packages/yak_utils/pubspec.yaml
@@ -12,3 +12,4 @@ dependencies:
 dev_dependencies:
   lints: ^2.0.0
   test: ^1.21.6
+  stub: ^1.0.0
